+++ src/jax_fdm/visualization/__init__.py
@@ -1,6 +1,7 @@
 from .artists import *  # noqa F403
 from .plotters import *  # noqa F403
 from .viewers import *  # noqa F403
+from .notebooks import *  # noqa F403
 
 
 __all__ = [name for name in dir() if not name.startswith('_')]