+++ src/jax_fdm/visualization/notebooks/__init__.py
@@ -1,0 +1,9 @@
+from .shapes import *  # noqa F403
+from .network_artist import *  # noqa F403
+from .viewer import *  # noqa F403
+from .register import register_artists
+
+
+register_artists()
+
+__all__ = [name for name in dir() if not name.startswith('_')]